+++ libtransmission/history.h
@@ -32,7 +32,7 @@ typedef struct tr_recentHistory tr_recentHistory;
  * @param precision how precise the history should be, in fractions of a second.
  *        For a precision of 1/20th of a second, use a precision of 20.
  */
-tr_recentHistory * tr_historyNew( int seconds, int precision );
+tr_recentHistory * tr_historyNew( unsigned int seconds, unsigned int precision );
 
 /** @brief destroy an existing tr_recentHistory object. */
 void tr_historyFree( tr_recentHistory * );
@@ -42,13 +42,13 @@ void tr_historyFree( tr_recentHistory * );
  * @param when the current time in msec, such as from tr_date()
  * @param n how many items to add to the history's counter
  */
-void tr_historyAdd( tr_recentHistory *, uint64_t when, double n );
+void tr_historyAdd( tr_recentHistory *, uint64_t when, unsigned int n );
 
 /**
  * @brief count how many events have occurred in the last N seconds.
  * @param when the current time in msec, such as from tr_date()
  * @param seconds how many seconds to count back through.
  */
-double tr_historyGet( const tr_recentHistory *, uint64_t when, int seconds );
+unsigned int tr_historyGet( const tr_recentHistory *, uint64_t when, unsigned int seconds );
 
 #endif
